requirements.txt
hg+http://bitbucket.org/pygame/pygame
Pillow==2.8.1
